updates user experience score
and test function. Takes a account_no, env_score, and it'll add the ENV score as XP to the user's "user_experience" score.

Please note i multiplied the env_score by 100, so we are adding full integers into our database instead of floats. This serves a double purpose of being nicer to store, and bigger numbers mean XP is more satisfying.